Refactor to base exportable code
