+++ external/anomaly/ote_anomalib/exportable_code/__init__.py
@@ -2,5 +2,6 @@
 
 from .anomaly_classification import AnomalyClassification
 from .anomaly_segmentation import AnomalySegmentation
+from .base import AnomalyBase
 
-__all__ = ["AnomalyClassification", "AnomalySegmentation"]
+__all__ = ["AnomalyClassification", "AnomalySegmentation", "AnomalyBase"]
